+++ src/journey.js
@@ -79,7 +79,7 @@
         _parameter_Document.body.appendChild( _element_Dialog );
 
         var closeButton = createElement( "button", "close" );
-        closeButton.onclick = hideDialog;
+        closeButton.onclick = onDialogClose;
         _element_Dialog.appendChild( closeButton );
 
         _element_Dialog_Title = createElement( "div", "title" );
@@ -98,6 +98,17 @@
         _element_Dialog_Next_Button = createElement( "button", "next" );
         _element_Dialog_Next_Button.onclick = onDialogNext;
         buttons.appendChild( _element_Dialog_Next_Button );
+    }
+
+    function onDialogClose() {
+        var bindingOptions = _elements_Attributes_Json[ _elements_Attributes_Keys[ _elements_Attributes_Position ] ];
+
+        if ( isDefined( bindingOptions.element ) ) {
+            fireCustomTrigger( bindingOptions.onClose, bindingOptions.element );
+        }
+
+        removeFocusClassFromLastElement( false );
+        hideDialog();
     }
 
     function onDialogPrevious() {
@@ -187,7 +198,9 @@
         _element_Dialog.style.display = "none";
     }
 
-    function removeFocusClassFromLastElement() {
+    function removeFocusClassFromLastElement( fireCustomTrigger ) {
+        fireCustomTrigger = isDefined( fireCustomTrigger ) ? fireCustomTrigger : true;
+
         var bindingOptions = _elements_Attributes_Json[ _elements_Attributes_Keys[ _elements_Attributes_Position ] ];
 
         if ( isDefined( bindingOptions.element ) ) {
@@ -197,7 +210,9 @@
                 bindingOptions.element.style.position = _element_Focus_Element_PositionStyle;
             }
 
-            fireCustomTrigger( bindingOptions.onLeave, bindingOptions.element );
+            if ( fireCustomTrigger ) {
+                fireCustomTrigger( bindingOptions.onLeave, bindingOptions.element );
+            }
         }
     }
 
@@ -289,6 +304,7 @@
     function buildAttributeOptionCustomTriggers( options ) {
         options.onEnter = getDefaultFunction( options.onEnter, null );
         options.onLeave = getDefaultFunction( options.onLeave, null );
+        options.onClose = getDefaultFunction( options.onClose, null );
 
         return options;
     }