+++ docs/CHANGE_LOG.md
@@ -2,6 +2,9 @@
 
 ## Version 0.2.0:
 - Minor documentation improvements.
+- Added a new binding custom trigger called "onClose", which states an event that should be triggered when the dialog is closed on a specific DOM element.
+- Fixed a fault that caused the CSS class "journey-js-element-focus" to remain on an element when the dialog is closed.
+- 
 
 <br>